chore: add database to assets

+++ build/download-node.js
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 var tar = require('tar-stream')
 var zlib = require('zlib')
-const { chmodSync, createWriteStream, mkdirSync, readFileSync, writeFileSync } = require('fs')
+const { chmodSync, createWriteStream, mkdirSync, readFileSync, writeFileSync, existsSync } = require('fs')
 const { join } = require('path')
 const request = require('request')
 const unzipper = require('unzipper')
@@ -29,47 +29,60 @@ const main = async (platform) => {
 
       switch (platform) {
          case 'darwin': {
-            var extract = tar.extract()
-            var chunks = []
+            const nodePath = join(assetDir, 'node')
 
-            extract.on('entry', function (header, stream, next) {
-               if (header.name == 'node-v14.16.0-darwin-x64/bin/node') {
-                  stream.on('data', function (chunk) {
-                     chunks.push(chunk)
+            // TODO: Detecting if node exists by filename only is a bit naive
+            if (existsSync(nodePath)){
+               var extract = tar.extract()
+               var chunks = []
+
+               extract.on('entry', function (header, stream, next) {
+                  if (header.name == 'node-v14.16.0-darwin-x64/bin/node') {
+                     stream.on('data', function (chunk) {
+                        chunks.push(chunk)
+                     })
+                  }
+
+                  stream.on('end', function () {
+                     next()
                   })
-               }
 
-               stream.on('end', function () {
-                  next()
+                  stream.resume()
                })
 
-               stream.resume()
-            })
+               extract.on('finish', function () {
+                  if (chunks.length) {
+                     var data = Buffer.concat(chunks)
+                     const name = nodePath
+                     writeFileSync(name, data)
+                     chmodSync(name, '755')
+                  }
+               })
 
-            extract.on('finish', function () {
-               if (chunks.length) {
-                  var data = Buffer.concat(chunks)
-                  const name = join(assetDir, 'node')
-                  writeFileSync(name, data)
-                  chmodSync(name, '755')
-               }
-            })
+               return request(url).pipe(zlib.createGunzip()).pipe(extract)
+            }
 
-            return request(url).pipe(zlib.createGunzip()).pipe(extract)
+            break
          }
          case 'win':
          default: {
-            const zip = request(url).pipe(unzipper.Parse({ forceStream: true }))
+            const nodePath = join(assetDir, 'node.exe')
+            // TODO: Detecting if node exists by filename only is a bit naive
+            if (existsSync(nodePath)){
+               const zip = request(url).pipe(unzipper.Parse({ forceStream: true }))
 
-            for await (const entry of zip) {
-               const fileName = entry.path
+               for await (const entry of zip) {
+                  const fileName = entry.path
 
-               if (fileName === `${base}/node.exe`) {
-                  entry.pipe(createWriteStream(join(assetDir, 'node.exe')))
-               } else {
-                  entry.autodrain()
+                  if (fileName === `${base}/node.exe`) {
+                     entry.pipe(createWriteStream(nodePath))
+                  } else {
+                     entry.autodrain()
+                  }
                }
             }
+
+            break
          }
       }
    } catch (e) {
@@ -79,5 +92,3 @@ const main = async (platform) => {
 
 main('darwin')
 main('win')
-
-// main(process.argv.slice(2)[0])
